chore: deduplicate msToString and bytesToString into isomorphic/formatUtils
Move duplicated time/size formatting utilities into a single shared
location at utils/isomorphic/formatUtils.ts and update all consumers.
