+++ packages/playwright-core/src/tools/trace/traceCli.ts
@@ -23,6 +23,7 @@ import { TraceModel, buildActionTree } from '../../utils/isomorphic/trace/traceM
 import { TraceLoader } from '../../utils/isomorphic/trace/traceLoader';
 import { renderTitleForCall } from '../../utils/isomorphic/protocolFormatter';
 import { asLocatorDescription } from '../../utils/isomorphic/locatorGenerators';
+import { msToString, bytesToString } from '../../utils/isomorphic/formatUtils';
 import { ZipTraceLoaderBackend } from './traceParser';
 
 import type { ActionTraceEventInContext } from '@isomorphic/trace/traceModel';
@@ -147,43 +148,6 @@ export async function loadTrace(traceFile: string): Promise<{ model: TraceModel,
 
 export async function loadTraceModel(traceFile: string): Promise<TraceModel> {
   return (await loadTrace(traceFile)).model;
-}
-
-function msToString(ms: number): string {
-  if (ms < 0 || !isFinite(ms))
-    return '-';
-  if (ms === 0)
-    return '0';
-  if (ms < 1000)
-    return ms.toFixed(0) + 'ms';
-  const seconds = ms / 1000;
-  if (seconds < 60)
-    return seconds.toFixed(1) + 's';
-  const minutes = seconds / 60;
-  if (minutes < 60)
-    return minutes.toFixed(1) + 'm';
-  const hours = minutes / 60;
-  if (hours < 24)
-    return hours.toFixed(1) + 'h';
-  const days = hours / 24;
-  return days.toFixed(1) + 'd';
-}
-
-function bytesToString(bytes: number): string {
-  if (bytes < 0 || !isFinite(bytes))
-    return '-';
-  if (bytes === 0)
-    return '0';
-  if (bytes < 1000)
-    return bytes.toFixed(0);
-  const kb = bytes / 1024;
-  if (kb < 1000)
-    return kb.toFixed(1) + 'K';
-  const mb = kb / 1024;
-  if (mb < 1000)
-    return mb.toFixed(1) + 'M';
-  const gb = mb / 1024;
-  return gb.toFixed(1) + 'G';
 }
 
 function formatTimestamp(ms: number, base: number): string {
