+++ packages/trace-viewer/src/ui/networkResourceDetails.tsx
@@ -25,7 +25,8 @@ import { CopyToClipboardTextButton } from './copyToClipboard';
 import { getAPIRequestCodeGen } from './codegen';
 import type { Language } from '@isomorphic/locatorGenerators';
 import { isJsonMimeType, isXmlMimeType } from '@isomorphic/mimeType';
-import { msToString, useAsyncMemo, useSetting } from '@web/uiUtils';
+import { useAsyncMemo, useSetting } from '@web/uiUtils';
+import { msToString } from '@isomorphic/formatUtils';
 import type { Entry } from '@trace/har';
 import { useTraceModel } from './traceModelContext';
 import { Expandable } from '@web/components/expandable';
