+++ packages/web/src/uiUtils.ts
@@ -16,6 +16,7 @@
 
 import React from 'react';
 
+
 // Recalculates the value when dependencies change.
 export function useAsyncMemo<T>(fn: () => Promise<T>, deps: React.DependencyList, initialValue: T, resetValue?: T) {
   const [value, setValue] = React.useState<T>(initialValue);
@@ -64,54 +65,6 @@ export function useMeasureForRef<T extends Element>(ref?: React.RefObject<T | nu
     };
   }, [recalculateMeasure, ref]);
   return [measure, recalculateMeasure];
-}
-
-export function msToString(ms: number): string {
-  if (ms < 0 || !isFinite(ms))
-    return '-';
-
-  if (ms === 0)
-    return '0';
-
-  if (ms < 1000)
-    return ms.toFixed(0) + 'ms';
-
-  const seconds = ms / 1000;
-  if (seconds < 60)
-    return seconds.toFixed(1) + 's';
-
-  const minutes = seconds / 60;
-  if (minutes < 60)
-    return minutes.toFixed(1) + 'm';
-
-  const hours = minutes / 60;
-  if (hours < 24)
-    return hours.toFixed(1) + 'h';
-
-  const days = hours / 24;
-  return days.toFixed(1) + 'd';
-}
-
-export function bytesToString(bytes: number): string {
-  if (bytes < 0 || !isFinite(bytes))
-    return '-';
-
-  if (bytes === 0)
-    return '0';
-
-  if (bytes < 1000)
-    return bytes.toFixed(0);
-
-  const kb = bytes / 1024;
-  if (kb < 1000)
-    return kb.toFixed(1) + 'K';
-
-  const mb = kb / 1024;
-  if (mb < 1000)
-    return mb.toFixed(1) + 'M';
-
-  const gb = mb / 1024;
-  return gb.toFixed(1) + 'G';
 }
 
 export function lowerBound<S, T>(array: S[], object: T, comparator: (object: T, b: S) => number, left?: number, right?: number): number {
